+++ annotations/nocedit.exe/pseudocode/src/core/set.cpp/CDemonSet_getReverbPresetAtPosition_FUN_0056fac0.cpp
@@ -2,11 +2,11 @@
 // Address: 0056fac0
 // Address Range: [[0056fac0, 0056fb7f]]
 // Convention: __cdecl
-// Signature: float __cdecl core_set_cpp_CDemonSet_getReverbPresetAtPosition_FUN_0056fac0(CDemonSet *this_ptr,CVector3f *position)
+// Signature: int __cdecl core_set_cpp_CDemonSet_getReverbPresetAtPosition_FUN_0056fac0(CDemonSet *this_ptr,CVector3f *position)
 
 #include "nocturne.h"
 
-float __cdecl core_set_cpp_CDemonSet_getReverbPresetAtPosition_FUN_0056fac0(CDemonSet *this_ptr,CVector3f *position)
+int __cdecl core_set_cpp_CDemonSet_getReverbPresetAtPosition_FUN_0056fac0(CDemonSet *this_ptr,CVector3f *position)
 
 {
   SRoom *pSVar1;
@@ -28,11 +28,11 @@ float __cdecl core_set_cpp_CDemonSet_getReverbPresetAtPosition_FUN_0056fac0(CDem
                 (this_ptr_00,&local_30,&local_24);
       if (((ABS(local_30.x) <= (pSVar1->extents).x) && (ABS(local_30.y) <= (pSVar1->extents).y)) &&
          (ABS(local_30.z) <= (pSVar1->extents).z)) {
-        return (float)pSVar1->reverb_size;
+        return pSVar1->reverb_size;
       }
       iVar2 = iVar2 + 1;
       this_ptr_00 = (CMatrix3x3f *)&this_ptr_00[1].m[2].z;
     } while (iVar2 < this_ptr->room_count);
   }
-  return (float)this_ptr->default_room_size;
+  return this_ptr->default_room_size;
 }
